Address my (ivo's) comments

+++ cpp/arcticdb/python/adapt_read_dataframe.hpp
@@ -42,7 +42,7 @@ inline py::tuple adapt_read_df(ReadResult&& ret, std::any* const handler_data) {
             pyuser_meta,
             multi_key_meta,
             std::move(node_results),
-            static_cast<int>(ret.sorted)
+            ret.sorted
     );
 };
